完成了mybatis实例练习，学会了关闭session，但是还是无法向MySQL插入数据，就好像项目mybatis操作的数据库ssm和cmd里面展示的数据库ssm不是一个库似的。。不明白为什么，得问问老蔡

+++ mybatis/p57_mybatis01/mybatis01/src/main/java/com/alison/mybatis/mapper/RoleMapper.java
@@ -8,11 +8,11 @@ public interface RoleMapper {
 	
 	public int insertRole(Role role);
 	
-	public int deleteRole(Long id);
+	public int deleteRole(int id);
 	
 	public int updateRole(Role role);
 	
-	public Role getRole(long id);
+	public Role getRole(int id);
 	
 	public List<Role> findRoles(String Rolename);
 }
